pyproject.toml
[tool.poetry]
name = "cybercare-test-task"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false 

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.8.1"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
django = "5.1.3"
djangorestframework = "3.15.2"
idna = "3.10"
requests = "2.32.3"
sqlparse = "0.5.2"
typing-extensions = "4.12.2"
urllib3 = "2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
